requirements.txt
# ════════════════════════════════════════════════════════════
# FLUX ACHAT PORTAL - Backend FastAPI
# ════════════════════════════════════════════════════════════

# Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Base de données
mysql-connector-python==8.3.0
sqlalchemy==2.0.25

# Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS
starlette==0.35.1

# Date/Time
python-dateutil==2.8.2
pydantic[email]
bcrypt